+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Public/BugsnagPerformance.cs
@@ -2,9 +2,11 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 using BugsnagNetworking;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using UnityEngine.Scripting;
 
 namespace BugsnagUnityPerformance
 {
@@ -27,6 +29,7 @@ namespace BugsnagUnityPerformance
         private PValueUpdater _pValueUpdater;
         private static List<Span> _potentiallyOpenSpans = new List<Span>();
         private Func<BugsnagNetworkRequestInfo, BugsnagNetworkRequestInfo> _networkRequestCallback;
+        private static string[] _tracePropagationUrlMatchPatterns;
 
         public static void Start(PerformanceConfiguration configuration)
         {
@@ -41,6 +44,10 @@ namespace BugsnagUnityPerformance
                     return;
                 }
                 IsStarted = true;
+            }
+            if (configuration.TracePropagationUrlMatchPatterns != null)
+            {
+                _tracePropagationUrlMatchPatterns = configuration.TracePropagationUrlMatchPatterns.ToArray();
             }
             ValidateApiKey(configuration.ApiKey);
             if (ReleaseStageEnabled(configuration))
@@ -241,24 +248,80 @@ namespace BugsnagUnityPerformance
         private void OnRequestSend(BugsnagUnityWebRequest request)
         {
             var url = request.url;
+            bool shouldCreateSpan = true;
             if (_networkRequestCallback != null)
             {
                 var callbackResult = _networkRequestCallback.Invoke(new BugsnagNetworkRequestInfo(url));
                 if (callbackResult == null || string.IsNullOrEmpty(callbackResult.Url))
                 {
-                    return;
+                    shouldCreateSpan = false;
                 }
 
                 url = callbackResult.Url;
             }
-            var span = _spanFactory.CreateAutomaticNetworkSpan(request, url);
-            lock (_networkSpansLock)
+
+            Span networkSpan = null;
+
+            if (shouldCreateSpan)
             {
-                _networkSpans[request] = span;
+                networkSpan = _spanFactory.CreateAutomaticNetworkSpan(request, url);
+                lock (_networkSpansLock)
+                {
+                    _networkSpans[request] = networkSpan;
+                }
+            }
+
+            if (ShouldAddTraceParentHeader(request.url))
+            {
+                string parentId = "";
+                string traceId = "";
+                bool sampled = false;
+
+                if (networkSpan != null)
+                {
+                    parentId = networkSpan.SpanId;
+                    traceId = networkSpan.TraceId;
+                    sampled = _sampler.Sampled(networkSpan,false);
+                }
+                else
+                {
+                    ISpanContext currentContext = _spanFactory.GetCurrentContext();
+                    if (currentContext != null)
+                    {
+                        parentId = currentContext.SpanId;
+                        traceId = currentContext.TraceId;
+                    }
+                }
+                if (string.IsNullOrEmpty(parentId))
+                {
+                    return;
+                }
+                request.SetRequestHeader("traceparent", BuildTraceParentHeader(traceId,parentId,sampled));
             }
         }
 
-       
+        private static bool ShouldAddTraceParentHeader(string url)
+        {
+            if (_tracePropagationUrlMatchPatterns == null || _tracePropagationUrlMatchPatterns.Length == 0)
+            {
+                return true;
+            }
+            foreach (var pattern in _tracePropagationUrlMatchPatterns)
+            {
+                if (Regex.IsMatch(url, pattern))
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        private static string BuildTraceParentHeader(string traceId, string parentSpanId, bool sampled)
+        {
+            return $"00-{traceId}-{parentSpanId}-{(sampled ? "01" : "00")}";
+        }
+
+
 
         private void OnRequestAbort(BugsnagUnityWebRequest request)
         {
@@ -299,6 +362,11 @@ namespace BugsnagUnityPerformance
             }
         }
 
+        public static ISpanContext GetCurrentSpanContext()
+        {
+            return _sharedInstance._spanFactory.GetCurrentContext();
+        }
+
         public static void ReportAppStarted()
         {
             AppStartHandler.ReportAppStarted();
@@ -319,6 +387,33 @@ namespace BugsnagUnityPerformance
                 }
             }
         }
+
+        [Serializable]
+        private class PerformanceState
+        {
+            public string currentContextSpanId;
+            public string currentContextTraceId;
+            public PerformanceState(string currentContextSpanId, string currentContextTraceId)
+            {
+                this.currentContextSpanId = currentContextSpanId;
+                this.currentContextTraceId = currentContextTraceId;
+            }
+        } 
+
+        [Preserve]
+        internal static string GetPerformanceState()
+        {
+            var context = GetCurrentSpanContext();
+            if (context != null)
+            {
+                var performanceState = new PerformanceState(context.SpanId, context.TraceId);
+                var json = JsonUtility.ToJson(performanceState);
+                return json;
+            }
+            return string.Empty;
+        }
+
+
 
     }
 }