+++ features/fixtures/mazerunner/Assets/Scripts/Scenarios/Configuration/BundleVersion.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 
 public class BundleVersion : Scenario
 {
-    public override void PrepareConfig(string apiKey, string host)
+    public override void PreparePerformanceConfig(string apiKey, string host)
     {
-        base.PrepareConfig(apiKey, host);
+        base.PreparePerformanceConfig(apiKey, host);
         SetMaxBatchSize(1);
         Configuration.BundleVersion = "1.2.3_BundleVersion";
     }