+++ features/fixtures/mazerunner/Assets/Scripts/Scenarios/Configuration/EnabledReleaseStages.cs
@@ -8,9 +8,9 @@ public class EnabledReleaseStages : Scenario
 
     private Span _span;
 
-    public override void PrepareConfig(string apiKey, string host)
+    public override void PreparePerformanceConfig(string apiKey, string host)
     {
-        base.PrepareConfig(apiKey, host);
+        base.PreparePerformanceConfig(apiKey, host);
         Configuration.ReleaseStage = "EnabledReleaseStages";
         Configuration.EnabledReleaseStages = new[] { "EnabledReleaseStages" };
         SetMaxBatchAgeSeconds(1);