+++ features/fixtures/mazerunner/Assets/Scripts/Scenarios/Network/NetworkCallbackUrlEdit.cs
@@ -8,9 +8,9 @@ using UnityEngine;
 public class NetworkCallbackUrlEdit : Scenario
 {
 
-    public override void PrepareConfig(string apiKey, string host)
+    public override void PreparePerformanceConfig(string apiKey, string host)
     {
-        base.PrepareConfig(apiKey, host);
+        base.PreparePerformanceConfig(apiKey, host);
         Configuration.NetworkRequestCallback = NetworkCallback;
         SetMaxBatchSize(1);
     }